Changed repo link to go from WSL -> Windows instead of Windows -> WSL

+++ windows-install.ps1
@@ -40,50 +40,60 @@ function Export-File {
 $REPOSITORY_ZIP_NAME = "environment-master-temp.zip"
 $REPOSITORY_PARENT_FOLDER = "environment-master-temp"
 $REPOSITORY_UNZIPED_NAME = "environment-master"
-$REPOSITORY_NAME = "environment"
+$REPOSITORY_NAME = "environment_temp"
 $REPOSITORY_BACKUP_NAME = "environment_backup"
 $SCRIPTS_FOLDER_NAME = "scripts"
 $WINDOWS_SETUP_SCRIPT_NAME = "windows-setup.ps1"
 $REPOSITORY_URL = "https://github.com/JateNensvold/environment/archive/master.zip"
 
-# Change path to current user directory
-Resolve-Path ~ | Set-Location
-$EXECUTION_DIRECTORY = Get-Location
-
 Write-Host "------------------------------------" -ForegroundColor Yellow
 Write-Host "Setting Location: " -NoNewline -ForegroundColor Green
 Write-Host $EXECUTION_DIRECTORY -ForegroundColor Cyan
 
-# Download Environment Git repo as a zip
-Write-Host "Downloading: " -ForegroundColor Green -NoNewline
-Write-Host $REPOSITORY_ZIP_NAME -ForegroundColor Cyan -NoNewline
-Write-Host " from: " -ForegroundColor Green -NoNewline
-Write-Host $REPOSITORY_URL -ForegroundColor Cyan
-Write-Host "------------------------------------" -ForegroundColor Yellow
-Invoke-WebRequest $REPOSITORY_URL -OutFile $REPOSITORY_ZIP_NAME
-
-
-# Unzip Git repo
-Expand-Archive $REPOSITORY_ZIP_NAME -DestinationPath $REPOSITORY_PARENT_FOLDER
-
-$REPOSITORY_UNZIPPED_PATH = Join-Path $REPOSITORY_PARENT_FOLDER $REPOSITORY_UNZIPED_NAME
-
-Export-File $REPOSITORY_NAME $REPOSITORY_BACKUP_NAME -Force
-mv $REPOSITORY_UNZIPPED_PATH $REPOSITORY_NAME
-rm $REPOSITORY_ZIP_NAME
-rm -r $REPOSITORY_PARENT_FOLDER
+# Change path to current user directory
+Resolve-Path ~ | Set-Location
+$EXECUTION_DIRECTORY = Get-Location
 
 # Need to use multiple Join-path for Powershell 5
 $REPOSITORY_ABSOLUTE_PATH = Join-Path $EXECUTION_DIRECTORY $REPOSITORY_NAME
 $WINDOWS_SETUP_SCRIPT_ABSOLUTE_PATH = (Join-Path $REPOSITORY_ABSOLUTE_PATH (
         Join-Path $SCRIPTS_FOLDER_NAME $WINDOWS_SETUP_SCRIPT_NAME))
 
-Write-Host "Unziped and moved repository to: " -ForegroundColor Green -NoNewline
-Write-Host $REPOSITORY_ABSOLUTE_PATH -ForegroundColor Cyan
+if (!Test-FilePath $REPOSITORY_ABSOLUTE_PATH) {
+
+    Write-Host "Repository already exists at: " -NoNewline -ForegroundColor Green
+    Write-Host $REPOSITORY_ABSOLUTE_PATH -ForegroundColor Cyan
+    Write-Host "Skipping Repository download stage..." -ForegroundColor Green
+
+}
+else {
+
+    # Download Environment Git repo as a zip
+    Write-Host "Downloading: " -ForegroundColor Green -NoNewline
+    Write-Host $REPOSITORY_ZIP_NAME -ForegroundColor Cyan -NoNewline
+    Write-Host " from: " -ForegroundColor Green -NoNewline
+    Write-Host $REPOSITORY_URL -ForegroundColor Cyan
+    Write-Host "------------------------------------" -ForegroundColor Yellow
+    Invoke-WebRequest $REPOSITORY_URL -OutFile $REPOSITORY_ZIP_NAME
+
+
+    # Unzip Git repo
+    Expand-Archive $REPOSITORY_ZIP_NAME -DestinationPath $REPOSITORY_PARENT_FOLDER
+
+    $REPOSITORY_UNZIPPED_PATH = Join-Path $REPOSITORY_PARENT_FOLDER $REPOSITORY_UNZIPED_NAME
+
+    Export-File $REPOSITORY_NAME $REPOSITORY_BACKUP_NAME -Force
+    mv $REPOSITORY_UNZIPPED_PATH $REPOSITORY_NAME
+    rm $REPOSITORY_ZIP_NAME
+    rm -r $REPOSITORY_PARENT_FOLDER
+
+    Write-Host "Unziped and moved repository to: " -ForegroundColor Green -NoNewline
+    Write-Host $REPOSITORY_ABSOLUTE_PATH -ForegroundColor Cyan
+}
+
+Write-Host "------------------------------------" -ForegroundColor Yellow
 Write-Host "Executing: " -ForegroundColor Green -NoNewline
 Write-Host $WINDOWS_SETUP_SCRIPT_ABSOLUTE_PATH -ForegroundColor Cyan
-Write-Host "------------------------------------" -ForegroundColor Yellow
-
 Write-Host "Running windows setup scripts..." -ForegroundColor Yellow
 
 & $WINDOWS_SETUP_SCRIPT_ABSOLUTE_PATH
